refactor: 설정 고급 섹션 제거 및 일반으로 통합
